@@ -18,9 +18,10 @@
     "gatsby-transformer-remark": "^2.2.6",
     "gatsby-transformer-sharp": "^2.1.8",
     "prop-types": "^15.7.2",
-    "react": "^16.6.3",
-    "react-dom": "^16.6.3",
-    "react-helmet": "^5.2.0"
+    "react": "16.8.6",
+    "react-dom": "16.8.6",
+    "react-helmet": "^5.2.0",
+    "sharp": "^0.22.0"
   },
   "keywords": [
     "gatsby"
@@ -42,6 +43,7 @@
     "eslint-plugin-node": "^8.0.1",
     "eslint-plugin-promise": "^4.0.1",
     "eslint-plugin-react": "^7.12.4",
+    "eslint-plugin-react-hooks": "^1.6.0",
     "eslint-plugin-standard": "^4.0.0",
     "husky": "^1.3.1",
     "prettier": "^1.15.2"
